rag_app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
pydantic==2.8.2
python-dotenv==1.0.1
orjson==3.10.7
Jinja2==3.1.4
azure-search-documents==11.6.0b9
azure-identity==1.17.1
openai==1.40.3
numpy==1.26.4
